+++ tests/__helpers__/auth.ts
@@ -1,18 +1,10 @@
 import { ApplicationType } from '../../src/services/auth/applications/types';
 
-export const applicationDataList = {
-  data: [
-    {
-      name: 'fitbit',
-    },
-  ],
-};
-
 export const newApplication = {
   id: '507f191e810c19729de860ea',
   type: ApplicationType.oauth2,
-  description: 'Test App',
   name: 'Test',
+  description: 'Test App',
   updateTimestamp: new Date('2019-04-15T11:23:19.670Z'),
   creationTimestamp: new Date('2019-04-15T11:23:19.670Z'),
 };
@@ -23,6 +15,11 @@ export const newApplicationVersion = {
   clientId: '0f4061a353c848eb0e02b80a2fe7bbc2254f1f77',
   clientSecret: '1f4221a223c848eb0e02b80a2fe7bbc2254f1f33',
   creationTimestamp: new Date('2019-04-15T11:23:19.670Z'),
+};
+
+export const applicationData = {
+  ...newApplication,
+  versions: [newApplicationVersion],
 };
 
 export const authorizationData = {