+++ CHANGELOG.md
@@ -10,10 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `exh.users.globalRoles` now has `find`, `findFirst`, `findById` and `findByName` methods
 - `exh.users.groupRoles` now has `find`, `findFirst`, `findById` and `findByName` methods
+- `exh.auth.applications` now has `find`, `findFirst`, `findById` and `findByName` methods
 
 ### Deprecated
 - `exh.users.globalRoles.get` is deprecated in favor of `exh.users.globalRoles.find`
 - `exh.users.groupRoles.get` is deprecated in favor of `exh.users.groupRoles.find`
+- `exh.auth.applications.get` is deprecated in favor of `exh.auth.applications.find`
 
 ### Fixed
 - The `data` for `exh.profiles.update` is now correctly typed.